data_orchestration/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
snowflake-connector-python 
python-dotenv
apache-airflow-providers-snowflake==5.0.0
